+++ client/app/styles/layout/_footer.scss
