+++ client/app/styles/layout/_header.scss
