+++ client/app/styles/pages/_home.scss
