+++ server/package.json
@@ -25,7 +25,7 @@
     "@nestjs/core": "^7.5.1",
     "@nestjs/platform-express": "^7.5.1",
     "@nestjs/swagger": "^4.7.12",
-    "class-validator": "^0.12.2",
+    "class-validator": "^0.13.1",
     "pg": "^8.5.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
