+++ server/src/models/posts/dto/post.dto.ts
@@ -1,0 +1,12 @@
+import { IsInt, IsString } from 'class-validator';
+
+export class PostDto {
+  @IsInt()
+  id: number;
+
+  @IsString()
+  title: string;
+
+  @IsString()
+  body?: string;
+}
