--- server/src/models/posts/interfaces/post.interface.ts
@@ -1,5 +1,0 @@
-export interface PostInterface {
-  id: number;
-  title: string;
-  body?: string;
-}
