+++ server/src/models/posts/post.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Post } from './entities/post.entity';
-import { PostInterface } from './interfaces/post.interface';
+// import { PostDto } from './dto/post.dto';
 
 @Injectable()
 export class PostService {
@@ -11,7 +11,7 @@ export class PostService {
     private postRepository: Repository<Post>,
   ) {}
 
-  async fetchAllPosts(): Promise<PostInterface[]> {
+  async fetchAllPosts(): Promise<Post[]> {
     return await this.postRepository.find();
   }
 }
