+++ server/src/models/users/dto/user.dto.ts
@@ -1,0 +1,12 @@
+import { IsInt, IsString } from 'class-validator';
+
+export class UserDto {
+  @IsInt()
+  readonly id: number;
+
+  @IsString()
+  readonly username: string;
+
+  @IsString()
+  readonly password: string;
+}
