--- server/src/models/users/interfaces/user.interface.ts
@@ -1,5 +1,0 @@
-export interface UserInterface {
-  id: number;
-  username: string;
-  password?: string;
-}
