+++ server/src/models/users/user.controller.ts
@@ -1,13 +1,16 @@
 import { Controller, Get } from '@nestjs/common';
-import { UserInterface } from './interfaces/user.interface';
+import { ApiTags } from '@nestjs/swagger';
+// import { UserDto } from './dto/user.dto';
+import { User } from './entities/user.entity';
 import { UserService } from './user.service';
 
 @Controller('user')
+@ApiTags('user')
 export class UserController {
   constructor(private userService: UserService) {}
 
   @Get()
-  fetchAllUsers(): Promise<UserInterface[]> {
+  fetchAllUsers(): Promise<User[]> {
     return this.userService.fetchAllUsers();
   }
 }
